advent/pyproject.toml
[tool.poetry]
name = "advent"
version = "0.1.0"
description = ""
authors = ["Dronakurl <konrad2710@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
ipython = "^8.18.1"
numpy = "^1.26.2"
pandas = "^2.1.3"
networkx = "^3.2.1"
matplotlib = "^3.8.2"
scipy = "^1.11.4"
pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[tool.pyright]
include = ["**/*.py"]
exclude = ["**/__pycache__"]
ignore = ["**/__init__.py"]
reportUnusedFunction = false
reportUnusedVariable = false
reportUnusedImport = false
useLibraryCodeForTypes = true
reportUnnecessaryTypeIgnoreComment = true
# typeCheckingMode = "off"
enableTypeIgnoreComments = true

[tool.ruff]
line-length = 120
exclude = ["**/__init__.py"]

[tool.pytest.ini_options]
python_functions = "test_*"
python_files = "test*.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
